feat: improve canvas generation and GUI feedback
- Generate one physical overlay per visible palette color and keep black empty.\n- Apply overlay RGB values to both Splitter_3 and the colored Part.\n- Center the canvas on X and raise it by 0.75 on Y.\n- Add an activated purple Gyro connected to Base point 1.\n- Add centered copy-enabled GUI notifications and center the main window.\n- Update asset paths, tests, and project documentation for the new build layout.

+++ src/rtg/cframe.py
@@ -183,7 +183,7 @@ class CFrame:
         return f"CFrame({x:.2f}, {y:.2f}, {z:.2f})"
 
 
-def create_pixel_offset_cframe(x: int, y: int, spacing: float = 1.0) -> CFrame:
+def create_pixel_offset_cframe(x: float, y: float, spacing: float = 1.0) -> CFrame:
     """
     Create a CFrame offset for placing a pixel at grid position (x, y).
